feat: Разделение монолита на микросервисы
      - выделены границы сервисов: events, requests, users
      - реализовано межсервисное взаимодействие через Feign + Eureka
      - подключён Config Server
      - настроен API Gateway
      - удалена вынесенная логика из main-service
      - подготовлена структура для Database per Service

+++ stats/stats-server/src/main/java/ru/practicum/mapper/StatisticsMapper.java
@@ -5,6 +5,6 @@ import ru.practicum.statistics.dto.ViewStatsDto;
 
 public class StatisticsMapper {
     public static ViewStatsDto toViewStatsDto(ViewStats viewStats) {
-        return new ViewStatsDto(viewStats.getApp(), viewStats.getUri(), viewStats.getHits());
+        return new ViewStatsDto(viewStats.getApp(), viewStats.getUri(), viewStats.getHits().intValue());
     }
 }
